Go back to a single environment, wrap with FormatObs
