+++ baselines/competition/competition_env.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2020. Huawei Technologies Co., Ltd. All rights reserved.
+# Copyright (C) 2022. Huawei Technologies Co., Ltd. All rights reserved.
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -18,132 +18,33 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
 
-import csv
 import logging
-import math
-import os
-from pathlib import Path
-from typing import Any, Dict, List, Optional, Sequence, Tuple, Union
-from uuid import uuid4
+from typing import Any, Dict, List, Optional, Sequence, Tuple
 
 import gym
-import gym.spaces as spaces
 import numpy as np
 from envision.client import Client as Envision
 from smarts.core import seed as smarts_seed
-from smarts.core.agent_interface import AgentInterface, AgentType
-from smarts.core.controllers import ActionSpaceType, ControllerOutOfLaneException
+from smarts.core.agent_interface import AgentInterface
+from smarts.core.controllers import ActionSpaceType
 from smarts.core.coordinates import Heading
 from smarts.core.scenario import Scenario
 from smarts.core.sensors import Observation
 from smarts.core.smarts import SMARTS
 from smarts.core.utils.logging import timeit
 from smarts.core.utils.math import vec_to_radians
+from smarts.env.wrappers.format_obs import FormatObs
+from smarts.zoo.agent_spec import AgentSpec
 
-NUM_WAYPOINTS = 5
 MAX_MPS = 100
 AGENT_ID = "EGO"
 
 
-def _filter(obs: Observation, target_position, env):
-    def _clip(formatted_obs, observation_space):
-        result = {}
-        for k, v in formatted_obs.items():
-            if type(observation_space[k]) == spaces.Tuple:
-                tuple_vals = []
-                for inner_val, inner_space in zip(v, observation_space[k]):
-                    tuple_vals.append(
-                        np.clip(inner_val, inner_space.low, inner_space.high)
-                    )
-                result[k] = tuple(tuple_vals)
-            else:
-                result[k] = np.clip(
-                    v, observation_space[k].low, observation_space[k].high
-                )
-        return result
-
-    obs = {
-        "position": obs.ego_vehicle_state.position,
-        "linear_velocity": obs.ego_vehicle_state.linear_velocity,
-        "target_position": target_position,
-        "rgb": obs.top_down_rgb.data.astype(np.uint8),
-        "waypoints": tuple(
-            [
-                np.hstack(
-                    (
-                        obs.waypoint_paths[0][i].pos,
-                        np.array(
-                            [
-                                float(obs.waypoint_paths[0][i].heading),
-                                obs.waypoint_paths[0][i].speed_limit,
-                            ]
-                        ),
-                    )
-                )
-                for i in range(NUM_WAYPOINTS)
-            ]
-        ),
-    }
-    obs = _clip(obs, env.observation_space)
-    assert env.observation_space.contains(
-        obs
-    ), "Observation mismatch with observation space. Less keys in observation space dictionary."
-    return obs
-
-
-class CompetitionEnv(gym.Env):
+class CompetitionBaseEnv(gym.Env):
     """A specific competition environment."""
 
     metadata = {"render.modes": ["human"]}
     """Metadata for gym's use"""
-    action_space = spaces.Box(low=-20.0, high=20.0, shape=(2,), dtype=np.float)
-    observation_space = spaces.Dict(
-        {
-            # position x, y, z in meters
-            "position": spaces.Box(
-                low=-math.inf,
-                high=math.inf,
-                shape=(3,),
-                dtype=np.float32,
-            ),
-            # Velocity
-            "linear_velocity": spaces.Box(
-                low=-MAX_MPS,
-                high=MAX_MPS,
-                shape=(3,),
-                dtype=np.float32,
-            ),
-            # target position x, y, z in meters
-            "target_position": spaces.Box(
-                low=-math.inf,
-                high=math.inf,
-                shape=(3,),
-                dtype=np.float32,
-            ),
-            # RGB image
-            "rgb": spaces.Box(
-                low=0,
-                high=255,
-                shape=(
-                    256,
-                    256,
-                    3,
-                ),
-                dtype=np.uint8,
-            ),
-            # nearest waypoints (2D position, heading, speed limit)
-            "waypoints": spaces.Tuple(
-                tuple(
-                    [
-                        spaces.Box(
-                            low=-math.inf, high=math.inf, shape=(4,), dtype=np.float32
-                        )
-                    ]
-                    * NUM_WAYPOINTS
-                )
-            ),
-        }
-    )
 
     def __init__(
         self,
@@ -186,6 +87,10 @@ class CompetitionEnv(gym.Env):
             waypoints=True,
             action=ActionSpaceType.TargetPose,
         )
+        agent_spec = AgentSpec(interface=agent_interface)
+
+        # Needs to be set for StdObs
+        self.agent_specs = {AGENT_ID: agent_spec}
 
         envision_client = None
         if not headless or envision_record_data_replay_path:
@@ -242,11 +147,6 @@ class CompetitionEnv(gym.Env):
             Tuple[Observation, float, bool, Any]:
                 Observation, reward, done, and info for the environment.
         """
-        assert self.action_space.contains(
-            np.array(agent_action)
-        ), f"Action {agent_action} must be within the action space: {self.action_space}"
-        observation, reward, done, extra = None, None, None, None
-
         heading = Heading(vec_to_radians(agent_action))  # naive heading
         l_p = self._last_obs.ego_vehicle_state.position
         target_pose = np.array(
@@ -263,21 +163,11 @@ class CompetitionEnv(gym.Env):
                 {AGENT_ID: target_pose}
             )
 
-        done = dones[AGENT_ID]
-        reward = rewards[AGENT_ID]
-        extra = {"score": extras["scores"][AGENT_ID], "env_obs": observations[AGENT_ID]}
         observation = observations[AGENT_ID]
-
         self._last_obs = observation
         self._current_time += observation.dt
-        target = [0, 0, 0]
 
-        return (
-            _filter(observation, target, self),
-            reward,
-            done,
-            extra,
-        )
+        return observations, rewards, dones, extras
 
     def reset(self) -> Observation:
         """Reset the environment and initialize to the next scenario.
@@ -294,9 +184,8 @@ class CompetitionEnv(gym.Env):
         self._last_obs = observation
 
         self._current_time += observation.dt
-        target = [0, 0, 0]
 
-        return _filter(observation, target, self)
+        return env_observations
 
     def render(self, mode="human"):
         """Does nothing."""
@@ -307,3 +196,81 @@ class CompetitionEnv(gym.Env):
         if self._smarts is not None:
             self._smarts.destroy()
             self._smarts = None
+
+
+class CompetitionEnv(gym.Env):
+    """A specific competition environment."""
+
+    def __init__(
+        self,
+        scenarios: Sequence[str],
+        headless: bool = True,
+        sim_name: Optional[str] = None,
+        max_episode_steps: Optional[int] = None,
+        seed: int = 42,
+        envision_endpoint: Optional[str] = None,
+        envision_record_data_replay_path: Optional[str] = None,
+    ):
+        """
+        Args:
+            sim_name (Optional[str], optional): Simulation name. Defaults to
+                None.
+            headless (bool, optional): If True, disables visualization in
+                Envision. Defaults to True.
+            seed (int, optional): Random number generator seed. Defaults to 42.
+            envision_endpoint (Optional[str], optional): Envision's uri.
+                Defaults to None.
+            envision_record_data_replay_path (Optional[str], optional):
+                Envision's data replay output directory. Defaults to None.
+            recorded_obs_path (Optional[Path, str], optional):
+                Output directory to write recorded observations to as a csv file.
+        """
+        base_env = CompetitionBaseEnv(
+            scenarios=scenarios,
+            headless=headless,
+            sim_name=sim_name,
+            max_episode_steps=max_episode_steps,
+            seed=seed,
+            envision_endpoint=envision_endpoint,
+            envision_record_data_replay_path=envision_record_data_replay_path,
+        )
+        self.env = FormatObs(env=base_env)
+
+    def step(
+        self, agent_action: Tuple[float, float]
+    ) -> Tuple[Observation, float, bool, Dict[str, Any]]:
+        """Steps the environment.
+
+        Args:
+            agent_action (Tuple[float, float]): Action taken by the agent.
+
+        Returns:
+            Tuple[Observation, float, bool, Any]:
+                Observation, reward, done, and info for the environment.
+        """
+        observations, rewards, dones, extras = self.env.step(agent_action)
+
+        done = dones[AGENT_ID]
+        reward = rewards[AGENT_ID]
+        extra = {"score": extras["scores"][AGENT_ID], "env_obs": observations[AGENT_ID]}
+        observation = observations[AGENT_ID]
+
+        return observation, reward, done, extra
+
+    def reset(self) -> Observation:
+        """Reset the environment and initialize to the next scenario.
+
+        Returns:
+            Observation: Agents' observation.
+        """
+        env_observations = self.env.reset()
+        observation = env_observations[AGENT_ID]
+        return observation
+
+    def render(self, mode="human"):
+        """Does nothing."""
+        pass
+
+    def close(self):
+        """Closes the environment and releases all resources."""
+        self.env.close()
